requirements.txt
alabaster==0.7.16
alembic==1.16.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
anywidget==0.9.18
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.4
bleach==6.2.0
blinker==1.9.0
boto3==1.38.24
botocore==1.38.24
bqplot==0.12.44
branca==0.8.1
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
colorlog==6.9.0
colour==0.1.5
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
DateTime==5.5
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
docutils==0.16
dotenv==0.9.9
duckdb==1.2.2
exceptiongroup==1.2.2
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.18.0
folium==0.19.5
fonttools==4.57.0
fqdn==1.5.1
futures==3.0.5
gdown==5.2.0
geojson==3.2.0
gitdb==4.0.12
GitPython==3.1.44
google==3.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.0rc1
google-api-python-client==2.170.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
google-genai==1.16.1
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
greenlet==3.2.3
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
imagesize==1.4.1
importlib_metadata==8.7.0
importlib_resources==6.5.2
ipyevents==2.0.2
ipyfilechooser==0.6.0
ipykernel==6.29.5
ipyleaflet==0.19.2
ipython==8.18.1
ipytree==0.2.2
ipyvue==1.11.2
ipyvuetify==1.11.1
ipywidgets==8.1.6
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.1
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter-events==0.12.0
jupyter-leaflet==0.19.2
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.1
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.14
kiwisolver==1.4.7
leafmap==0.43.11
llvmlite==0.43.0
logging==0.4.9.6
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.4
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.3
narwhals==1.37.1
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.4.1
notebook_shim==0.2.4
numba==0.60.0
numdifftools==0.9.41
numpy==2.0.2
nvidia-nccl-cu12==2.27.5
optuna==4.4.0
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
patsy==1.0.1
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.7
plotly==6.0.1
prometheus_client==0.21.1
prompt_toolkit==3.0.51
proto-plus==1.26.1
protobuf==5.29.4
psutil==7.0.0
psycopg2-binary==2.9.10
psygnal==0.12.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
pydeck==0.9.1
pyflux==0.3.1
Pygments==2.19.1
pyparsing==3.2.3
pyshp==2.3.1
PySocks==1.7.1
pystac==1.10.1
pystac-client==0.8.3
python-box==7.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==3.3.0
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
referencing==0.36.2
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.0.0
rpds-py==0.24.0
rsa==4.9.1
s3==3.0.0
s3transfer==0.13.0
scikit-learn==1.6.1
scipy==1.13.1
scooby==0.10.1
seaborn==0.13.2
Send2Trash==1.8.3
six==1.17.0
skforecast==0.16.0
smmap==5.0.2
sniffio==1.3.1
snowballstemmer==3.0.1
soupsieve==2.7
Sphinx==5.3.0
sphinx-rtd-theme==0.5.2
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
SQLAlchemy==2.0.41
stack-data==0.6.3
statsmodels==0.14.4
streamlit==1.45.0
tenacity==9.1.2
terminado==0.18.1
threadpoolctl==3.6.0
tinycss2==1.4.0
toml==0.10.2
tomli==2.2.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
traittypes==0.2.1
types-python-dateutil==2.9.0.20241206
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
uri-template==1.3.0
uritemplate==4.1.1
urllib3==1.26.20
watchdog==6.0.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
whitebox==2.3.6
whiteboxgui==2.3.0
widgetsnbextension==4.0.14
xgboost==2.1.4
xmltodict==0.14.2
xyzservices==2025.4.0
zipp==3.21.0
zope.interface==7.2